Add an option to allow self signed certificates
This change impacts the database sync feature when using a HTTPS location
as the database sharing location.

If the host where the database was being shared used a self-signed
certificate it was necessary to generate a new keystore with that hosts
certificate and copy this keystore into your JRE lib/security dir.

With this change an option was introduced to simply trust all hosts even
if they have an invalid X.509 certificate.