Set the DB dirty shortly after sync
Create a thread that will mark the database dirty after
a short period. Without this the database would remain
in a synced state until the user makes a change. The
longer we wait before syncing up the greater chance there
is that we'll miss changes made elsewhere and end up
with a conflicting version of the database.